pyproject.toml
[project]
name = "fontina"
version = "0.0.1"
description = "Visual font recognition (VFR) library"
authors = [{ name = "Alessio Placitelli", email = "a.placitelli@a2p.it" }]
dependencies = [
    "albumentations==1.3.1",
    "trdg==1.8.0",
    # We need to force the pillow version because of the use
    # of removed APIs in trdg 1.8.0 (e.g. `ImageFont.getsize`).
    "pillow==9.5.0",
    "pytorch-lightning[extra]==2.0.8",
    "pyyaml==6.0.1",
    "rstr==3.2.1",
    # TODO: On Windows, this must be manual.
    # pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu117
    "torch==2.0.1",
    "torchvision==0.15.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
linting = [
    "black==23.7.0",
    "flake8==6.1.0",
    "flake8-bugbear==23.7.10",
    "mypy==1.4.1",
    "pytest==7.4.0",
    "types-PyYAML==6.0.1",
    "types-pillow==9.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 90
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | assets
  | build
  | dist
)/
'''

[tool.mypy]
exclude = "(assets|build)/$"

[[tool.mypy.overrides]]
module = ["albumentations.*", "torchvision.*", "trdg.generators.*"]
ignore_missing_imports = true
